Login screen/feature
Created login screen and the possibility to login to the Cropool service. If the credentials are valid, user is redirected to the provisional HomeActivity which will show user's data about routes etc. in the future. HomeActivity should also redirect the user back to the login page if the tokens aren't valid. Created string resources for input field errors. PROBLEM: Sending plain-text passwords for the login process. If passwords are hashed, there is no way to compare them to the hashed passwords from the database. If we do not check the passwords at the CropoolAPI side, we can't know whether to issue the tokens or not. Work on #4, #5, #8, #9.