DateTime chat, margins, fix
Added date and time note when clicking on messages, adjusted the margins between messages. I think I fixed the 'not attached to activity' error (I replaced ConversationListFragment with ConversationListActivity). Still to implement: search. Creating new chats will be possible when a route is found and signed up to. Work on #16.