Profile picture update, FB sign in moved
Profile picture update implemented, waiting progress bar while uploading to FB storage. FB sign in was moved to HomeActivity in order to be able to access the current user's data from all other activities/fragments. Bug when showing the progile picture on MyAccountFragment fixed. Closes #13. #13 may be reopened when routes are implemented to the project.